Perform coding phase inspection

+++ src/main/java/drinkshop/service/StocService.java
@@ -67,10 +67,11 @@ public class StocService {
             double ramas = necesar;
 
             for (Stoc s : ingredienteStoc) {
-                if (ramas <= 0) break;
+                if (ramas <= 0)
+                    break;
 
                 double deScazut = Math.min(s.getCantitate(), ramas);
-                s.setCantitate((int)(s.getCantitate() - deScazut));
+                s.setCantitate(s.getCantitate() - deScazut);
                 ramas -= deScazut;
 
                 stocRepo.update(s);
